Don't disconnect from memcache on every request
Currently, Django disconnects from the cache on at the end of every HTTP request. This is a significant performance problem as it mangifies the number of round trip times (each connection requires a VERSION command, and if authenticating a LIST_MECHS and an AUTH command before any real work can be done) and incurs the cost of TCP startup every time.

There has been a long and slow push to get this fixed upstream in Django (https://code.djangoproject.com/ticket/11331, https://github.com/django/django/pull/4866), but it's a simple enough fix to do at the higher level library.